Deleted double tabs
When I copy-pasted one of this examples from the docs as a starting point, they were not accepted until I halved the amount of tabs so I figured that I'd remove them where they came from too. 

Is this correct? And if so, please chack: did I remove ALL the excess tabs correctly?