include where clause checking if a field is empty
This issue came up on the irc the other day - 
nmarques attempted to use {% setcontent members = 'staff' where {position: "!''"} %} to return records where staff was not empty, but found it didn't work. Instead, nmarques figured out that '!' worked: {% setcontent members = 'staff' where {position: '!'} %} 

Since this seems to me to be an unexpected behavior, I figured it should be documented. (Alternatively if it's a bug, then an issue might be opened?)