Correcting the instructions to modify extension.
The instructions to modify the starter extension omitted several important steps, such as the need to rename the starter extension class and to update the bolt-class setting in composer.json. Without these steps, when new developers follow this tutorial, the extension fails to install and/or Bolt labels it as invalid.